requirements.txt
# Env
python-dotenv>=1.0.0

# API
fastapi>=0.115.0
uvicorn>=0.30.0

# HTTP Client (async)
httpx>=0.27.0

# Google Gemini
google-genai>=1.0.0

# Formatação
black>=24.0.0

# Testes
pytest>=8.0.0
pytest-asyncio>=1.0.0
